requirements.txt
google-generativeai
requests
streamlit
python-dotenv
